src: cleanup msvc warnings.

+++ test/Utility/HttpServerMock.h
@@ -35,8 +35,8 @@ class HttpServerMock : public cloudstorage::IHttpServer {
     MOCK_CONST_METHOD0(method, std::string());                         // NOLINT
     MOCK_CONST_METHOD0(url, std::string());                            // NOLINT
     MOCK_CONST_METHOD4(mocked_response,
-                       IResponse::Pointer(int, const IResponse::Headers&, int,
-                                          IResponse::ICallback*));
+                       IResponse::Pointer(int, const IResponse::Headers&,
+                                          int64_t, IResponse::ICallback*));
     IResponse::Pointer response(
         int code, const IResponse::Headers& headers, int64_t size,
         IResponse::ICallback::Pointer cb) const override {
